Reorganize automated generation code
