+++ src/mappings/ro-to-wikidata.sssom.tsv
@@ -6,21 +6,23 @@
 #mapping_provider: http://obofoundry.org/ontology/ro
 #subject_source: http://purl.obolibrary.org/obo/ro.owl
 subject_id	subject_label	predicate_id	object_id	object_label	match_type	comments
-RO:0001025		owl:equivalentProperty	wd:P276	location	Curated	derived from wikidata via SPARQL
-RO:0002211		owl:equivalentProperty	wd:P128	regulates (molecular biology)	Curated	derived from wikidata via SPARQL
-BFO:0000051		owl:equivalentProperty	wd:P527	has part	Curated	derived from wikidata via SPARQL
-BFO:0000050		owl:equivalentProperty	wd:P361	part of	Curated	derived from wikidata via SPARQL
-RO:0002204		owl:equivalentProperty	wd:P702	encoded by	Curated	derived from wikidata via SPARQL
-RO:0002205		owl:equivalentProperty	wd:P688	encodes	Curated	derived from wikidata via SPARQL
-RO:0002411		owl:equivalentProperty	wd:P828	has cause	Curated	derived from wikidata via SPARQL
-RO:0002162		owl:equivalentProperty	wd:P703	found in taxon	Curated	derived from wikidata via SPARQL
-RO:0002404		owl:equivalentProperty	wd:P1542	has effect	Curated	derived from wikidata via SPARQL
-RO:0002405		owl:equivalentProperty	wd:P1536	immediate cause of	Curated	derived from wikidata via SPARQL
-RO:0003001		owl:equivalentProperty	wd:P2849	produced by	Curated	derived from wikidata via SPARQL
-RO:0002302		owl:equivalentProperty	wd:P2176	drug used for treatment	Curated	derived from wikidata via SPARQL
-RO:0002008		owl:equivalentProperty	wd:P1382	partially coincident with	Curated	derived from wikidata via SPARQL
-RO:0002005		owl:equivalentProperty	wd:P3189	innervated by	Curated	derived from wikidata via SPARQL
-RO:0002202		owl:equivalentProperty	wd:P3094	develops from	Curated	derived from wikidata via SPARQL
-RO:0000087		owl:equivalentProperty	wd:P2868	subject has role	Curated	derived from wikidata via SPARQL
-RO:0002379		owl:equivalentProperty	wd:P3403	coextensive with	Curated	derived from wikidata via SPARQL
-RO:0002134		owl:equivalentProperty	wd:P3190	innervates	Curated	derived from wikidata via SPARQL
+BFO:0000050	part of	owl:equivalentProperty	wd:P361	part of	Curated	derived from wikidata via SPARQL
+BFO:0000051	has part	owl:equivalentProperty	wd:P527	has part	Curated	derived from wikidata via SPARQL
+RO:0000087	has role	owl:equivalentProperty	wd:P2868	subject has role	Curated	derived from wikidata via SPARQL
+RO:0001025	located in	owl:equivalentProperty	wd:P276	location	Curated	derived from wikidata via SPARQL
+RO:0002005	innervated_by	owl:equivalentProperty	wd:P3189	innervated by	Curated	derived from wikidata via SPARQL
+RO:0002008	coincident with	owl:equivalentProperty	wd:P1382	partially coincident with	Curated	derived from wikidata via SPARQL
+RO:0002134	innervates	owl:equivalentProperty	wd:P3190	innervates	Curated	derived from wikidata via SPARQL
+RO:0002162	in taxon	owl:equivalentProperty	wd:P703	found in taxon	Curated	derived from wikidata via SPARQL
+RO:0002202	develops from	owl:equivalentProperty	wd:P3094	develops from	Curated	derived from wikidata via SPARQL
+RO:0002204	gene product of	owl:equivalentProperty	wd:P702	encoded by	Curated	derived from wikidata via SPARQL
+RO:0002205	has gene product	owl:equivalentProperty	wd:P688	encodes	Curated	derived from wikidata via SPARQL
+RO:0002211	regulates	owl:equivalentProperty	wd:P128	regulates (molecular biology)	Curated	derived from wikidata via SPARQL
+RO:0002302	is treated by substance	owl:equivalentProperty	wd:P2176	drug used for treatment	Curated	derived from wikidata via SPARQL
+RO:0002379	spatially coextensive with	owl:equivalentProperty	wd:P3403	coextensive with	Curated	derived from wikidata via SPARQL
+RO:0002404	causally downstream of	owl:equivalentProperty	wd:P1542	has effect	Curated	derived from wikidata via SPARQL
+RO:0002405	immediately causally downstream of	owl:equivalentProperty	wd:P1536	immediate cause of	Curated	derived from wikidata via SPARQL
+RO:0002411	causally upstream of	owl:equivalentProperty	wd:P828	has cause	Curated	derived from wikidata via SPARQL
+RO:0003001	produced by	owl:equivalentProperty	wd:P2849	produced by	Curated	derived from wikidata via SPARQL
+RO:0012001	has small molecule activator	owl:equivalentProperty	wd:P3771	activator of	Curated	derived from wikidata via SPARQL
+RO:0012002	has small molecule inhibitor	owl:equivalentProperty	wd:P3776	inhibitor of	Curated	derived from wikidata via SPARQL
